% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RK4ODECpp}
\alias{RK4ODECpp}
\title{The fourth-order Runger-Kutta method using Cpp}
\usage{
RK4ODECpp(fun, a, b, n, y0)
}
\arguments{
\item{fun}{the ordinary differential equations}

\item{a}{the upper bound of the interval}

\item{b}{the lower bound of the interval}

\item{n}{the length of the interval}

\item{y0}{initial value}
}
\value{
The solution of the equation on the interval [a,b], the data type is dataframe.
}
\description{
This model is used to calculate the initial values of ordinary differential equations
}
\examples{
\dontrun{
fun <- function(x,y){
x-y+1
}
a=0
b=1
n=10
y0=1
library(Rcpp)
sourceCpp(paste0('D:/R/Rwork/Statistics_computing/RK.cpp'))
RK4ODECpp = RK4ODECpp(fun, a, b, n, y0)
}
}
