% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RombergCpp}
\alias{RombergCpp}
\title{The Romberg quadrature method using Cpp}
\usage{
RombergCpp(f, a, b, TOL)
}
\arguments{
\item{f}{the function}

\item{a}{the upper bound of the interval}

\item{b}{the lower bound of the interval}

\item{TOL}{integration error limit}
}
\value{
A list is returned, with the first value s being the integral result, 
the second value R being the iterative process, and the third value k being the number of iterations
}
\description{
This method is used to solve numerical integrals
}
\examples{
\dontrun{
f <- function(x) exp(x)*x^2+exp(x)*sin(x)+exp(x)*cos(x)+2**x  # 积分函数
a <- 0  # 积分下限
b <- 1  # 积分上限
TOL <- 1e-16  # 误差容限
resultCpp <- RombergCpp(f, a, b, TOL)
}
}
