\name{SA24204125-package}
\alias{SA24204125-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  Some functions are provided in this package.
}
\author{
Yingsi Liang

Maintainer: Yingsi Liang <1759878480@qq.com>
}
\keyword{ package }
